+++ src/main/java/org/WaialuaRobotics359/robot/subsystems/Wrist.java
@@ -42,7 +42,7 @@ public class Wrist extends SubsystemBase {
     }
 
     public double GetPosition() {
-        return mWristMotor.getSelectedSensorPosition();
+        return (int) mWristMotor.getSelectedSensorPosition();
      }
 
      public double GetPositionInches(){
